requirements.txt
apispec==3.3.0
attrs==19.3.0
backpack==0.1
blinker==1.4
cleo==0.6.0
click==7.1.1
clikit==0.5.1
Faker==0.7.18
flasgger==0.9.4
Flask==1.1.2
flask-apispec==0.8.8
Flask-JWT-Extended==3.24.1
flask-orator==0.2.0
Flask-WTF==0.14.3
importlib-metadata==1.6.0
inflection==0.3.1
itsdangerous==1.1.0
Jinja2==2.11.1
jsonschema==3.2.0
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
marshmallow==3.5.1
mistune==0.8.4
mysqlclient==2.0.1
orator==0.9.7
pastel==0.2.0
pendulum==1.5.1
psutil==5.7.0
pyaml==16.12.2
Pygments==2.6.1
PyJWT==1.7.1
pylev==1.3.0
pyrsistent==0.16.0
python-dateutil==2.8.1
pytz==2019.3
pytzdata==2019.3
PyYAML==5.3.1
simplejson==3.17.0
six==1.14.0
text-unidecode==1.2
tzlocal==1.5.1
webargs==5.5.3
Werkzeug==1.0.1
woops==0.2.1
wrapt==1.12.1
WTForms==2.2.1
WTForms-JSON==0.3.3
zipp==3.1.0
python-dotenv==0.14.0